requirements.txt
requests
numpy
Pillow
scikit-image
scikit-learn
mysql-connector-python
flask
flask_sqlalchemy
flask_cors
gunicorn

